Lazily initialize variables on parameter parsing
When we query for the version, we don't want to traverse the whole
directory. This leads to '-V' returning after a long time if
executed in the home directory. Factor it out to a method.

We also don't need the other variables, like config parsing to be
initialized for 'version' or 'help', move them to the if block.